+++ src/main/webapp/myPageActionGo.jsp
@@ -6,8 +6,8 @@
 <%@ page import = "dao.*" %>
 <%@ page import="java.net.URLEncoder"%>
 <%
-	String idIn = (String)(session.getAttribute("loginId"));
-	MallDaoLee mallDaoLee = new MallDaoLee();
+String idIn = (String)(session.getAttribute("loginId"));
+	UserDao userDao = new UserDao();
 	
 	String updatePw = request.getParameter("updatePw");
 	String updateName = request.getParameter("updateName");
@@ -16,7 +16,7 @@
 	
 	
 	// 마이페이지 패스워드 체크 성공 이후 업데이트
-	String reslut = mallDaoLee.updateUser(idIn, updatePw, updateName, updatePhone, updateAddress);
+	String reslut = userDao.updateUser(idIn, updatePw, updateName, updatePhone, updateAddress);
 	System.out.println(reslut); 
 	if (reslut.equals("false")){	// 업데이트 실패시
 		String msg = URLEncoder.encode("ERROR : 0000 (관리자에게 문의하세요)");
@@ -25,7 +25,4 @@
 	}else if(reslut.equals("true")){	// 업데이트 성공
 		response.sendRedirect(request.getContextPath()+"/main.jsp");
 	}
-
-
-
 %>
